+++ rosidl_generator_cpp/resource/idl__builder.hpp.em
@@ -25,7 +25,7 @@ include_directives = set()
 #define @(header_guard_variable)
 
 #include "@(include_base)__struct.hpp"
-#include <rosidl_generator_cpp/message_initialization.hpp>
+#include <rosidl_runtime_cpp/message_initialization.hpp>
 #include <algorithm>
 #include <utility>
