+++ rosidl_generator_cpp/resource/idl__traits.hpp.em
@@ -25,7 +25,7 @@ include_directives = set()
 #define @(header_guard_variable)
 
 #include "@(include_base)__struct.hpp"
-#include <rosidl_generator_cpp/traits.hpp>
+#include <rosidl_runtime_cpp/traits.hpp>
 #include <stdint.h>
 #include <type_traits>
