+++ rosidl_typesupport_introspection_cpp/include/rosidl_typesupport_introspection_cpp/message_introspection.hpp
@@ -20,7 +20,7 @@
 
 #include "rosidl_generator_c/message_type_support_struct.h"
 
-#include "rosidl_generator_cpp/message_initialization.hpp"
+#include "rosidl_runtime_cpp/message_initialization.hpp"
 
 #include "rosidl_typesupport_introspection_cpp/visibility_control.h"
 
@@ -51,7 +51,7 @@ typedef struct ROSIDL_TYPESUPPORT_INTROSPECTION_CPP_PUBLIC MessageMembers
   uint32_t member_count_;
   size_t size_of_;
   const MessageMember * members_;
-  void (* init_function)(void *, rosidl_generator_cpp::MessageInitialization);
+  void (* init_function)(void *, rosidl_runtime_cpp::MessageInitialization);
   void (* fini_function)(void *);
 } MessageMembers;
